# docs/api/dnj-operations.openapi.yaml
openapi: 3.1.0
info:
  title: DNJ 2K26 Operations API
  version: 1.0.0-draft
  description: |
    Contract for the Next homologation operations API. It is the handoff
    contract for the future Go/PostgreSQL service; endpoints use a signed
    HTTP-only manager or admin cookie, never participant bearer tokens.
    Participant QR, Moment and special-event banner contracts are maintained
    alongside this document in dnj-experience.openapi.yaml.
servers:
  - url: /api
...
paths:
...
  /manager/session:
    get:
      {
        tags: [Manager session],
        summary: Current manager and one assigned scope,
        responses:
          {
            "200": { description: Session },
            "401": { description: Unauthenticated },
          },
      }
    post:
      tags: [Manager session]
      summary: Login and route to the account's assigned operation
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/ManagerLogin" } },
            },
        }
      responses:
        {
          "200": { description: Cookie session },
          "401": { description: Invalid credentials },
        }
    delete:
      {
        tags: [Manager session],
        summary: Logout,
        responses: { "200": { description: Session cleared } },
      }
...
  /manager/actions/runs:
    post:
      tags: [Radicalidade]
      summary: Create a game run and its first dynamic QR
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      required: [gameId],
                      properties: { gameId: { type: string, format: uuid } },
                    },
                },
            },
        }
      responses:
        {
          "201":
            {
              description: Run and one-time QR,
              content:
                {
                  application/json:
                    { schema: { $ref: "#/components/schemas/QrEnvelope" } },
                },
            },
        }
...
  /v1/activity-runs/current:
    get:
      tags: [Radicalidade]
      summary: Returns the current scanned Radicalidade run for the authenticated participant, including terminal state when `runId` is supplied
      parameters:
        [
          {
            name: runId,
            in: query,
            required: false,
            schema: { type: string, format: uuid },
          },
        ]
      responses:
        {
          "200": { description: Run state },
          "204": { description: No open run for this participant },
        }
...
  /manager/actions/pause:
    post:
      {
        tags: [Radicalidade],
        summary: Pause a running game,
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/RunIdRequest" } },
              },
          },
        responses: { "200": { description: Run paused } },
      }
...
  /manager/actions/results:
    post:
      tags: [Radicalidade]
      summary: Finalize 1st/2nd/3rd/participation scoring idempotently
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [runId, results]
              properties:
                runId: { type: string, format: uuid }
                results:
                  type: array
                  items:
                    type: object
                    required: [participantId, result]
                    properties:
                      participantId: { type: string, format: uuid }
                      result:
                        {
                          type: string,
                          enum: [first, second, third, participation],
                        }
      responses:
        {
          "200": { description: Points finalized },
          "409": { description: Run cannot be finalized },
        }
  /manager/actions/close:
    post:
      {
        tags: [Radicalidade],
        summary: Cancel or close a run,
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/RunIdRequest" } },
              },
          },
        responses: { "200": { description: Run closed } },
      }
...
  /admin/managers:
    get:
      {
        tags: [Admin],
        summary: List credentialed manager accounts and scopes,
        responses:
          {
            "200": { description: Managers },
            "401": { description: Admin required },
          },
      }
    post:
      {
        tags: [Admin],
        summary: Create or update credentialed manager/scope,
        responses:
          {
            "201": { description: Manager persisted },
            "400": { description: Invalid manager data },
          },
      }
    patch:
      {
        tags: [Admin],
        summary: Activate or deactivate a manager account,
        responses: { "200": { description: Manager updated } },
      }
...
components:
...
  schemas:
    ManagerLogin:
      {
        type: object,
        required: [email, password],
        properties:
          {
            email: { type: string, format: email },
            password: { type: string, format: password },
          },
      }
...
    RunIdRequest:
      {
        type: object,
        required: [runId],
        properties: { runId: { type: string, format: uuid } },
      }
...
    QrEnvelope:
      {
        type: object,
        required: [qrId, qrPayload, qrImageUrl, expiresAt],
        properties:
          {
            qrId: { type: string, format: uuid },
            qrPayload:
              {
                type: string,
                description: Returned only to the operator at generation time,
              },
            qrImageUrl:
              {
                type: string,
                format: uri,
                description: Renderable QR image,
                possibly a data URI,
              },
            expiresAt: { type: string, format: date-time },
          },
      }